pyproject.toml
[tool.poetry]
name = "poetry install"
version = "0.1.0"
description = "poetry shell"
authors = ["poetry init"]
license = "exit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.37.1"
pandas = "^2.2.2"
fpdf = "^1.7.2"
reportlab = "^4.2.2"
plotly = "^5.23.0"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
